+++ Editor/AddressableImportRule.cs
@@ -12,40 +12,56 @@ public enum AddressableImportRuleMatchType
     /// *, matches any number of characters
     /// ?, matches a single character
     /// </summary>
+    [Tooltip("Simple wildcard.\n\"*\" matches any number of characters.\n\"?\" matches a single character.")]
     Wildcard = 0,
 
     /// <summary>
     /// Regex pattern
     /// </summary>
+    [Tooltip("A regular expression pattern.")]
     Regex
+}
+
+public enum LabelWriteMode
+{
+    Add,
+    Replace
 }
 
 [System.Serializable]
 public class AddressableImportRule
 {
     /// <summary>
-    /// Path pattern
+    /// Path pattern.
     /// </summary>
+    [Tooltip("The assets in this path will be processed.")]
     public string path;
 
     /// <summary>
-    /// Path match type
+    /// Method used to parse the Path.
     /// </summary>
+    [Tooltip("The path parsing method.")]
     public AddressableImportRuleMatchType matchType;
 
     /// <summary>
-    /// Addressable group name
+    /// The group the asset will be added.
     /// </summary>
-    [Tooltip("Leaves blank for the default group.")]
+    [Tooltip("The group name in which the Addressable will be added. Leave blank for the default group.")]
     public string groupName;
 
     /// <summary>
-    /// Label reference list
+    /// Defines if labels will be added or replaced.
     /// </summary>
+    public LabelWriteMode LabelMode;
+
+    /// <summary>
+    /// Label reference list.
+    /// </summary>
+    [Tooltip("The list of labels to be added to the Addressable Asset")]
     public List<AssetLabelReference> labelRefs;
 
     /// <summary>
-    /// Simplify address
+    /// Simplify address.
     /// </summary>
     [Tooltip("Simplify address to filename without extension.")]
     public bool simplified;
@@ -98,6 +114,67 @@ public class AddressableImportRule
                     yield return labelRef.labelString;
                 }
             }
+        }
+    }
+
+    public string ParseRegexPath(string assetPath, string replacement)
+    {
+        return AddressableImportRegex.ParsePath(assetPath, replacement);
+    }
+
+    static class AddressableImportRegex
+    {
+        const string pathregex = @"\%PATH\%\[\-{0,1}\d{1,3}\]"; // ie: $PATH$[0]
+
+        static public string[] GetPathArray(string path)
+        {
+            return path.Split('/');
+        }
+
+        static public string GetPathAtArray(string path, int idx)
+        {
+            return GetPathArray(path)[idx];
+        }
+
+        static public string ParsePath(string assetPath, string targetGroupName)
+        {
+            var _path = assetPath;
+            int i = 0;
+            var slashSplit = _path.Split('/');
+            var len = slashSplit.Length - 1;
+            var matches = Regex.Matches(targetGroupName, pathregex);
+            string[] parsedMatches = new string[matches.Count];
+            foreach (var match in matches)
+            {
+                string v = match.ToString();
+                var sidx = v.IndexOf('[') + 1;
+                var eidx = v.IndexOf(']');
+                int idx = int.Parse(v.Substring(sidx, eidx - sidx));
+                while (idx > len)
+                {
+                    idx -= len;
+                }
+                while (idx < 0)
+                {
+                    idx += len;
+                }
+                //idx = Mathf.Clamp(idx, 0, slashSplit.Length - 1);
+                parsedMatches[i++] = GetPathAtArray(_path, idx);
+            }
+
+            i = 0;
+            var splitpath = Regex.Split(targetGroupName, pathregex);
+            string finalPath = string.Empty;
+            foreach (var split in splitpath)
+            {
+                finalPath += splitpath[i];
+                if (i < parsedMatches.Length)
+                {
+                    finalPath += parsedMatches[i];
+                }
+                i++;
+            }
+            return finalPath;
         }
     }
 }