+++ CHANGELOG.md
@@ -4,8 +4,13 @@ All notable changes to this package will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## [0.2.0] - 2019-07-19
+## [0.2.0] - 2019-08-04
  - Added address replacement, working with regex rule to build the address based on information captured from the path.
+ - Added group replacement (dynamic named group i.e `%PATH%[0]`).
+ - Added option to automatically create groups if not exist.
+ - Added option to remove empty Asset Groups.
+ - Added option to define if labels from ruleset are added or replace the current ones.
+ - Improved documentation.
 
 ## [0.1.1] - 2019-07-19
  - Updated to addressable v1.1.5.